Remove client authorization (scope) stub and implement it
One of the results of this and the previous commit is that the client
is no longer indexed by a user/client id combo, but each client now has
as single client entry in the client map. Instead the scope granted to
the client is in a separate client authorization stored in the user record.